+++ crypto/pem/pem_lib.c
@@ -102,7 +102,7 @@ int PEM_def_callback(char *buf, int num, int w, void *key)
         i = EVP_read_pw_string_min(buf, MIN_LENGTH, num, prompt, w);
         if (i != 0) {
             PEMerr(PEM_F_PEM_DEF_CALLBACK, PEM_R_PROBLEMS_GETTING_PASSWORD);
-            memset(buf, 0, (unsigned int)num);
+            memset(buf, 0, num);
             return (-1);
         }
         j = strlen(buf);
